update figs and tables
