+++ haicosystem/utils/__init__.py
@@ -1,4 +1,5 @@
+from .aggregate_eval import get_avg_reward
 from .render import render_for_humans, rich_rendering
 from .sampler import BridgeSampler
 
-__all__ = ["render_for_humans", "rich_rendering", "BridgeSampler"]
+__all__ = ["render_for_humans", "rich_rendering", "BridgeSampler", "get_avg_reward"]
